fix(class-council): #EVAL-339 move skills validated percentage in class council feature on CompetenceNoteService
