+++ src/main/java/fr/openent/competences/service/BilanPeriodiqueService.java
@@ -1,9 +1,7 @@
 package fr.openent.competences.service;
 
 import fr.openent.competences.model.Service;
-import fr.openent.competences.model.achievements.AchievementsProgress;
 import fr.wseduc.webutils.Either;
-import io.vertx.core.Future;
 import io.vertx.core.Handler;
 import io.vertx.core.json.JsonArray;
 
@@ -43,18 +41,4 @@ public interface BilanPeriodiqueService {
     void getBilanPeriodiqueDomaineForGraph(final String idEleve, final String idEtablissement,
                                            final String idClasse, final Integer typeClasse, final String idPeriodeString,
                                            final Handler<Either<String, JsonArray>> handler);
-
-
-    /**
-     * Get skills validated percentage by subject for a student
-     *
-     * @param structureId structure identifier filter on
-     * @param studentId   student identifier to filter on
-     * @param periodId    period identifier to filter on
-     * @param groupId     group identifier to filter on
-     * @return return Future containing student SubjectSkillsValidatedPercentage inside his Achievements
-     */
-    Future<AchievementsProgress> getSubjectSkillsValidatedPercentage(String structureId, String studentId,
-                                                                     Long periodId, String groupId);
-
 }
